chore(http): add flags
